Remove worldwide priorities from features
As Worldwide Priorities are no longer able to be made through the admin
or visible as metadata, this commit removes all the Feature tests which
tested this code.

An exception is with the `translated_content_steps.rb` in which this had
to be changed to another format which can be translated.